Refactor cart.tsx, add updateCartItemQuantity function
This refactor of cart.tsx adds the updateCartItemQuantity function to
the CartProvider. This function allows for updating the quantity of
items in the cart. The addition of this function improves the
flexibility and user experience of the cart functionality. No issues are
associated with this change.

Note: Please replace "[issue reference]" with the relevant issue or omit
if not applicable.